% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_result.R
\name{get_one_result}
\alias{get_one_result}
\title{Get one survey from formr}
\usage{
get_one_result(run, survey, items = "")
}
\description{
\code{get_one_result} is for internal use of \code{get_results}
}
