% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_results.R
\name{get_raw_results}
\alias{get_raw_results}
\title{Get raw data from formr}
\usage{
get_raw_results(request = NULL, token = NULL)
}
\description{
\code{get_raw_results} is for internal use of \code{get_one_result} wich is in turn for internl use of \code{get_results}
}
