% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Get results from formr}
\usage{
get_results(run, surveys, items = NULL, do_prettify = TRUE)
}
\arguments{
\item{run}{specify wich formr run should be used (must be one)}

\item{items}{currently not used (don't change to something different then null)}

\item{do_prettify}{should the result be pretty? If you need more controll just call prettify explicitly}

\item{survey}{specify wich surveys from the specified run should be downloaded (can be more the one in one character vector)}
}
\description{
\code{get_results} returns the data of one or more surveys from one formr run
}
\seealso{
\code{\link{prettify}},\code{\link{formr_api_access_token}}
}
