% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_to_na_atomic.R
\name{null_to_na_atomic}
\alias{null_to_na_atomic}
\title{Turn null into na (for a vector)}
\usage{
null_to_na_atomic(x)
}
\description{
\code{null_to_na_atomic} is for internal use of get_results \code{null_to_na}
}
