% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Make a List of Surveys from formr pretty}
\usage{
prettify(list_of_surveys, conv_null = TRUE, conv_num = TRUE, join = T)
}
\arguments{
\item{list_of_surveys}{List of surveys from formr}

\item{conv_null}{turn null to na (formr gives you by default Null instead of NA)}

\item{conv_num}{convert should be numerics to numerics}

\item{join}{should a full join be done, see also \code{\link{join}}}
}
\description{
\code{prettify} returns the data of one or more surveys from one formr run
}
\seealso{
\code{\link{join}}
}
